feat: Support mixed MSRV in `--version-range`
This expands on the approach taken in #212 bucketing packages into
rust-versions to run.  If we skipped the MSRV (due to `--version-step`),
we automatically inject it.  If a package's MSRV isn't within the range,
we skip it.

Benefits
- Relatively simple to implement and to explain
- We keep the number of runs to a minimum by walking in lock-step the
  `--version-step`, independent of what each package' MSRV

I did have to specialize `--rust-version` vs `--version-range` to avoid
`--rust-version` range users walking more than they needed.

To keep the progress total accurate, I shifted the calculating of the
total from `determine_package_list` to after we have bucketed
everything.  To make this feasible, I saved off the how many iterations
a package will have without the version range being taken into account.

As a byproduct, this fixes a bug in #212 where it didn't take the
rust-version into account when determining the total.

Fixes #199